==> 04.sample3.Rmd <==
```{r function section, echo=FALSE,include=FALSE}

source("00.import.R")

```

---
title: "Single cell transcriptomic analysis on primary canine triple-negative mammary tumor cells infected by an oncolytic vaccinia virus to identify candidate resistance genes."
author: "K.Lebrigand"
output:
  BiocStyle::html_document:
    code_folding: hide
    number_sections: yes
    toc: yes  
---

# sample 3 sain

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=10}

data <- Read10X("data/s3/")
s3.sain <- CreateSeuratObject(data)
s3.sain[['sample']] <- "s3.sain"

mito.genes <- c('ATP6','COX1','COX3','CYTB','MT-ATP8','MT-CO2','MT-ND1','MT-ND3','MT-ND4','MT-ND6','ND2','ND4L','ND5')
dropouts <- Matrix::colSums(s3.sain@assays$RNA@data == 0)/nrow(s3.sain@assays$RNA)
ribo.genes <- grep(pattern = "^RP[SL]", x = rownames(s3.sain@assays$RNA), value = TRUE)
percent.mito <- Matrix::colSums(s3.sain@assays$RNA[mito.genes, ])/Matrix::colSums(s3.sain@assays$RNA)
percent.ribo <- Matrix::colSums(s3.sain@assays$RNA[ribo.genes, ])/Matrix::colSums(s3.sain@assays$RNA)
s3.sain[['percent.mito']] <- percent.mito
s3.sain[['percent.ribo']] <- percent.ribo
s3.sain[['dropouts']] <- dropouts
VlnPlot(s3.sain, features = c("nFeature_RNA", "nCount_RNA","dropouts","percent.ribo","percent.mito"), ncol=5, cols = "lightsteelblue3")
dim(s3.sain@assays$RNA)
s3.sain <- subset(s3.sain, subset = percent.mito < 0.25)
dim(s3.sain@assays$RNA)
VlnPlot(s3.sain, features = c("nFeature_RNA", "nCount_RNA","dropouts","percent.ribo","percent.mito"), ncol=5, cols = "lightsteelblue3")

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=8}

s3.sain <- SCTransform(object = s3.sain, verbose = FALSE)

s3.sain <- RunPCA(s3.sain, features = VariableFeatures(s3.sain), verbose = FALSE)
ElbowPlot(object = s3.sain)

s3.sain <- RunUMAP(object = s3.sain, dims = 1:30)
s3.sain <- FindNeighbors(object = s3.sain, do.plot=TRUE, dims = 1:30)
s3.sain <- FindClusters(object = s3.sain, resolution=0.2)
DimPlot(object = s3.sain, reduction = "umap")

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=14}

s3.sain <- CellCycleScoring(s3.sain, s.features = s.genes, g2m.features = g2m.genes, set.ident = FALSE)
FeaturePlot(s3.sain, features=c("S.Score","G2M.Score"))

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=8, fig.width=10}

markers <- FindAllMarkers(object = s3.sain, only.pos = TRUE, min.pct = 0.25, thresh.use = 0.25)
top10 <- markers %>% group_by(cluster) %>% top_n(10, avg_logFC)
DoHeatmap(s3.sain, features=top10$gene, size=3.5)

```

# sample3 sain re-labeling, save .rds

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=8}

new.cluster.ids <- c("SPP1","MET","COL1A2","KRT14")
names(x = new.cluster.ids) <- levels(x = s3.sain)
s3.sain <- RenameIdents(object = s3.sain, new.cluster.ids)

my_levels <- c("SPP1","MET","COL1A2","KRT14")
# Relevel object@ident
s3.sain@active.ident <- factor(x = s3.sain@active.ident, levels = my_levels)

DimPlot(object = s3.sain, cols = cell_type_color, label = FALSE)

saveRDS(s3.sain, "output/s3.sain.rds")

freq_table <- prop.table(x = table(s3.sain@active.ident))*100
df <- as.data.frame(freq_table)
df$sample <- "sain"

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=8, fig.width=10}

markers <- FindAllMarkers(object = s3.sain, only.pos = TRUE, min.pct = 0.25, thresh.use = 0.25)
top10 <- markers %>% group_by(cluster) %>% top_n(10, avg_logFC)
DoHeatmap(s3.sain, features=top10$gene, size=3.5)
write.table(markers, file="output/s3.sain.csv", sep=",")

```

# sample3 infected

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=10}

data <- Read10X("data/s3_infected/")
s3.inf <- CreateSeuratObject(data)
s3.inf[['sample']] <- "s3.inf"

genes.vv <- genes.vv[genes.vv %in% rownames(s3.inf@assays$RNA@counts)]
percent.vv <- Matrix::colSums(s3.inf@assays$RNA@counts[genes.vv, ])/Matrix::colSums(s3.inf@assays$RNA)

genes.early <- genes.early[genes.early %in% rownames(s3.inf@assays$RNA@counts)]
percent.early <- Matrix::colSums(s3.inf@assays$RNA@counts[genes.early, ])/Matrix::colSums(s3.inf@assays$RNA@counts)

genes.intermediate <- genes.intermediate[genes.intermediate %in% rownames(s3.inf@assays$RNA@counts)]
percent.intermediate <- Matrix::colSums(s3.inf@assays$RNA@counts[genes.intermediate, ])/Matrix::colSums(s3.inf@assays$RNA@counts)

genes.late <- genes.late[genes.late %in% rownames(s3.inf@assays$RNA@counts)]
percent.late <- Matrix::colSums(s3.inf@assays$RNA@counts[genes.late, ])/Matrix::colSums(s3.inf@assays$RNA@counts)

# remove vv genes and re-do analysis
iv = match(genes.vv, rownames(data))
all(genes.vv==rownames(data)[iv])
virus <- data[iv,]
dog_only <- data[-iv,]

s3.inf <- CreateSeuratObject(dog_only)
s3.inf[['percent.vv']] <- percent.vv
s3.inf[['percent.early']] <- percent.early
s3.inf[['percent.intermediate']] <- percent.intermediate
s3.inf[['percent.late']] <- percent.late
s3.inf[['color']] <- percent.late

s3.inf[['color']][which(s3.inf@meta.data$percent.vv < 0.01),] <- "#9593a2"
s3.inf[['color']][which(s3.inf@meta.data$percent.vv > 0.01),] <- "#297c88"
s3.inf[['color']][which(s3.inf@meta.data$percent.late > 0.02),] <- "#ed4700"
table(s3.inf@meta.data$color)

mito.genes <- c('ATP6','COX1','COX3','CYTB','MT-ATP8','MT-CO2','MT-ND1','MT-ND3','MT-ND4','MT-ND6','ND2','ND4L','ND5')
dropouts <- Matrix::colSums(s3.inf@assays$RNA@data == 0)/nrow(s3.inf@assays$RNA)
ribo.genes <- grep(pattern = "^RP[SL]", x = rownames(s3.inf@assays$RNA), value = TRUE)
percent.mito <- Matrix::colSums(s3.inf@assays$RNA[mito.genes, ])/Matrix::colSums(s3.inf@assays$RNA)
percent.ribo <- Matrix::colSums(s3.inf@assays$RNA[ribo.genes, ])/Matrix::colSums(s3.inf@assays$RNA)
s3.inf[['percent.mito']] <- percent.mito
s3.inf[['percent.ribo']] <- percent.ribo
s3.inf[['dropouts']] <- dropouts

VlnPlot(s3.inf, features = c("nFeature_RNA", "nCount_RNA","dropouts","percent.ribo","percent.mito"), ncol=5, cols = "lightsteelblue3")
dim(s3.inf@assays$RNA)

s3.inf[["VIRUS"]] <- CreateAssayObject(counts = virus)
DefaultAssay(object = s3.inf) <- "VIRUS"
s3.inf <- NormalizeData(s3.inf, normalization.method = "LogNormalize", scale.factor = 1000)
s3.inf <- ScaleData(s3.inf)

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=4, fig.width=6}

ggplot(data=s3.inf@meta.data, aes(x=percent.vv,y=percent.mito)) +  geom_point(shape = 21, colour = "black", fill = s3.inf@meta.data$color) +
      theme_bw() + theme(panel.grid.minor=element_blank(), panel.grid.major=element_blank()) +
      ggtitle("percent.mito = f(percent.vv)") +
#      geom_abline(slope=1, intercept=-15, linetype=3) +
      labs(x="percent.vv",y="percent.mito")

DefaultAssay(object = s3.inf) <- "RNA"
s3.inf <- subset(s3.inf, subset = percent.mito < 0.25)
dim(s3.inf)

ggplot(data=s3.inf@meta.data, aes(x=percent.late,y=percent.early)) +  geom_point(shape = 21, colour = "black", fill = s3.inf@meta.data$color) +
      theme_bw() + theme(panel.grid.minor=element_blank(), panel.grid.major=element_blank()) +
      ggtitle("percent.early = f(percent.late)") +
#      geom_abline(slope=1, intercept=-15, linetype=3) +
      labs(x="percent.late",y="percent.early")

ggplot(data=s3.inf@meta.data, aes(x=percent.late,y=percent.intermediate)) +  geom_point(shape = 21, colour = "black", fill = s3.inf@meta.data$color) +
      theme_bw() + theme(panel.grid.minor=element_blank(), panel.grid.major=element_blank()) +
      ggtitle("percent.intermediate = f(percent.late)") +
      labs(x="percent.late",y="percent.intermediate")

ggplot(data=s3.inf@meta.data, aes(x=log(nCount_VIRUS,10),y=log(nCount_RNA,10))) +  geom_point(shape = 21, colour = "black", fill = s3.inf@meta.data$color) +
      theme_bw() + theme(panel.grid.minor=element_blank(), panel.grid.major=element_blank()) +
      ggtitle("Host nCount_RNA = f(nCount_VIRUS)") +
      labs(x="nCount_VIRUS",y="nCount_RNA")

ggplot(data=s3.inf@meta.data, aes(x=percent.vv,y=nFeature_RNA)) +  geom_point(shape = 21, colour = "black", fill = s3.inf@meta.data$color) +
      theme_bw() + theme(panel.grid.minor=element_blank(), panel.grid.major=element_blank()) +
      ggtitle("nFeature_RNA = f(percent.vv)") +
#      geom_abline(slope=1, intercept=-15, linetype=3) +
      labs(x="percent.vv",y="nFeature_RNA")


```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=8}

s3.inf <- SCTransform(object = s3.inf, verbose = FALSE)
s3.inf <- RunPCA(s3.inf, features = VariableFeatures(s3.inf), verbose = FALSE)
ElbowPlot(object = s3.inf)

s3.inf <- RunUMAP(object = s3.inf, dims = 1:30)
s3.inf <- FindNeighbors(object = s3.inf, do.plot=TRUE, dims = 1:30)
s3.inf <- FindClusters(object = s3.inf, resolution=0.2)
DimPlot(object = s3.inf, label=TRUE)

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=14}

s3.inf <- CellCycleScoring(s3.inf, s.features = s.genes, g2m.features = g2m.genes)
FeaturePlot(s3.inf, features=c("S.Score","G2M.Score"))
FeaturePlot(s3.inf, features=c("percent.early","percent.late"))
FeaturePlot(s3.inf, features=c("KRT14"))
VlnPlot(s3.inf,c("SPP1","CAV2","COL1A2","KRT14"), ncol = 4)
VlnPlot(s3.inf,c("percent.early","percent.late"), ncol = 4)

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=8, fig.width=10}

markers <- FindAllMarkers(object = s3.inf, only.pos = TRUE, min.pct = 0.25, thresh.use = 0.25)
top10 <- markers %>% group_by(cluster) %>% top_n(8, avg_logFC)
DoHeatmap(s3.inf, features=top10$gene, size=3.5)

```

# sample3 infected re-labeling, save .rds

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=6, fig.width=8}

new.cluster.ids <- c("SPP1","MET","COL1A2","MDH1")
names(x = new.cluster.ids) <- levels(x = s3.inf)
s3.inf <- RenameIdents(object = s3.inf, new.cluster.ids)

my_levels <- c("SPP1","MET","COL1A2","MDH1")
# Relevel object@ident
s3.inf@active.ident <- factor(x = s3.inf@active.ident, levels = my_levels)

DimPlot(object = s3.inf, cols = cell_type_color, label = FALSE)

saveRDS(s3.inf, "output/s3.inf.rds")

freq_table <- prop.table(x = table(s3.inf@active.ident))*100
df2 <- as.data.frame(freq_table)
df2$sample <- "infected"

dd <- rbind(df,df2)
ggplot(dd, aes(sample, y=Freq, fill=Var1)) + 
    geom_bar(stat="identity") + 
    geom_label(aes(label = round(Freq,digits=1)), position = position_stack(vjust = 0.5), size = 3)

dd

```

```{r message=FALSE, warning=FALSE, echo=FALSE, eval=T, fig.height=8, fig.width=10}

markers <- FindAllMarkers(object = s3.inf, only.pos = TRUE, min.pct = 0.25, thresh.use = 0.25)
top10 <- markers %>% group_by(cluster) %>% top_n(8, avg_logFC)
DoHeatmap(s3.inf, features=top10$gene, size=3.5)

write.table(markers, file="output/s3.inf.csv", sep=",")

```
